requirements.txt
flask>=2.0.1
qrcode>=7.4.2
Pillow>=9.5.0
gunicorn>=20.1.0
